refactor: enhance output formatting and improve memory management in main functions

+++ CPP/CPP04/ex01/srcs/Brain.cpp
@@ -6,7 +6,7 @@
 /*   By: achauvie <achauvie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/15 09:47:09 by achauvie          #+#    #+#             */
-/*   Updated: 2026/05/15 09:54:35 by achauvie         ###   ########.fr       */
+/*   Updated: 2026/05/15 11:36:56 by achauvie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ Brain &Brain::operator=(const Brain &other)
 {
 	if (this != &other)
 	{
-		for (size_t i = 0; i < 100; i++)
+		for (int i = 0; i < 100; i++)
 			_ideas[i] = other._ideas[i];
 	}
 	std::cout << "Brain has been assigned!" << std::endl;
